add s21_sin() with tests

+++ include/s21_math.h
@@ -24,7 +24,7 @@ double s21_floor(double x);
 double s21_fmod(double x, double y);
 double s21_log(double x);
 long double s21_pow(double base, double exp);
-long double s21_sin(double x);
+double s21_sin(double x);
 double s21_sqrt(double x);
 long double s21_tan(double x);
 long double s21_trunc(double num);
